final few tests need fixing
