--- tests/test_commands.py
@@ -1,12 +1,0 @@
-import commands
-
-
-def test_get_usage(monkeypatch):
-    monkeypatch.setattr(commands, "top_level_command_args", {
-        "something1": None,
-        "something2": None,
-        "something3": None
-    })
-
-    result = commands.get_usage()
-    assert result == "usage: spotilistcli {help,configure,something1,something2,something3}"
