+++ tests/test_playlist.py
@@ -1,14 +1,19 @@
 import threading
 import pytest
 import tempfile
+
+from click.testing import CliRunner
+
 import sptfy
 import spotipy
 import json
 import os
 import ytmusic
 from datetime import datetime
+import commands
 from commands import playlist
-
+from commands.playlist.download import get_filename, preload_youtube_url_cache, add_youtube_url_to_songs, \
+    get_playlist_with_songs
 
 FROZEN_TIME = datetime(2024, 4, 3, 21, 2, 0)
 
@@ -37,7 +42,7 @@ def patch_datetime_now(monkeypatch):
             return FROZEN_TIME
 
     monkeypatch.setattr(
-        "commands.playlist.datetime",
+        "commands.playlist.download.datetime",
         MockedDateTime
     )
 
@@ -80,7 +85,7 @@ def test_get_filename(monkeypatch, sptfy_mock, patch_datetime_now):
         lambda self: expected_query_return
     )
 
-    actual = playlist.get_filename(sptfy_mock)
+    actual = get_filename(sptfy_mock)
 
     assert actual == "playlists-01234567890-2024_04_03T21_02_00.json"
 
@@ -102,7 +107,7 @@ def test_get_playlist_with_songs(monkeypatch, sptfy_mock):
         spotify_playlist_url="https://example.invalid"
     )
 
-    actual = playlist.get_playlist_with_songs(playlist_no_song, sptfy_mock)
+    actual = get_playlist_with_songs(playlist_no_song, sptfy_mock)
 
     assert actual.total == 347
 
@@ -146,6 +151,8 @@ def test_get_playlist_with_songs(monkeypatch, sptfy_mock):
 ])
 def test_show_playlist(monkeypatch, capfd, sptfy_mock,
                        args_list, playlist_found, expected_outputs, exit_expected, expected_exit_code):
+    runner = CliRunner()
+
     with open("tests/files/sptfy_playlist_items_queries.json.test", "r") as file:
         data = json.load(file)
     data_iter = iter(data)
@@ -160,17 +167,14 @@ def test_show_playlist(monkeypatch, capfd, sptfy_mock,
         lambda self, playlist_id: playlist_found
     )
     if exit_expected:
-        with pytest.raises(SystemExit) as exit_wrapper:
-            playlist.show_playlist(args_list, sptfy_mock)
-        assert exit_wrapper.type == SystemExit
-        assert exit_wrapper.value.code == expected_exit_code
+        result = runner.invoke(commands.show, args_list)
+        assert result.exit_code == expected_exit_code
     else:
-        playlist.show_playlist(args_list, sptfy_mock)
-        out, err = capfd.readouterr()
-        num_lines = len(out.split("\n"))
+        result = runner.invoke(commands.show, args_list)
+        num_lines = len(result.stdout.split("\n"))
         assert num_lines == 151
         for expected_output in expected_outputs:
-            assert expected_output in out
+            assert expected_output in result.stdout
 
 
 @pytest.mark.parametrize("args_list, expected_outputs", [
@@ -184,25 +188,26 @@ def test_show_playlist(monkeypatch, capfd, sptfy_mock,
     ),
 ])
 def test_list_playlists(monkeypatch, capfd, sptfy_mock, args_list, expected_outputs):
+    runner = CliRunner()
+
     with open("tests/files/sptfy_current_user_playlists_queries.json.test", "r") as file:
         data = json.load(file)
     data_iter = iter(data)
 
     monkeypatch.setattr(spotipy.Spotify, "current_user_playlists", lambda self, limit, offset: next(data_iter))
 
-    playlist.list_playlists(args_list, sptfy_mock)
-    out, err = capfd.readouterr()
-    num_lines = len(out.split("\n"))
+    result = runner.invoke(commands.list, args_list)
+    num_lines = len(result.stdout.split("\n"))
     assert num_lines == 119
     for expected_output in expected_outputs:
-        assert expected_output in out
+        assert expected_output in result.stdout
 
 
 def test_download_playlists_fail_filename(sptfy_mock):
     args_list = ["--filename", "does_not_end_in_dot_json"]
 
     with pytest.raises(SystemExit) as exit_wrapper:
-        playlist.download_playlists(args_list, sptfy_mock)
+        commands.download(args_list, sptfy_mock)
     assert exit_wrapper.type == SystemExit
     assert exit_wrapper.value.code == 1
 
@@ -251,7 +256,7 @@ def patch_get_playlist_with_songs(monkeypatch):
     )
 
     monkeypatch.setattr(
-        playlist, "get_playlist_with_songs",
+        commands.playlist, "get_playlist_with_songs",
         lambda playlist_with_no_song, sptfy_arg: playlist_with_songs
     )
 
@@ -275,7 +280,7 @@ def test_download_playlists(monkeypatch, capfd, sptfy_mock,
     temp_file = os.path.join(tmpdir, "test.json")
     args_list = ["--filename", temp_file]
 
-    playlist.download_playlists(args_list, sptfy_mock)
+    commands.download(args_list, sptfy_mock)
 
     with open(temp_file) as file:
         actual = json.load(file)
@@ -358,7 +363,7 @@ def test_add_youtube_url_to_songs(monkeypatch, ytm_mock, test_case):
     monkeypatch.setattr(ytmusic.YTM, "get_youtube_url", lambda self, song: "https://music.youtube.com/watch?v=wxyz")
     monkeypatch.setattr(threading.Event, "is_set", lambda self: False)
     event = threading.Event()
-    playlist.add_youtube_url_to_songs(playlist_input, ytm_mock, event)
+    add_youtube_url_to_songs(playlist_input, ytm_mock, event)
 
     assert playlist_input == expected
 
@@ -372,7 +377,7 @@ def test_download_playlists_with_youtube_url(monkeypatch, sptfy_mock,
 
     monkeypatch.setattr(ytmusic.YTM, "get_youtube_url", lambda self, song: "https://music.youtube.com/watch?v=wxyz")
 
-    playlist.download_playlists(args_list, sptfy_mock)
+    commands.download(args_list, sptfy_mock)
 
     with open(temp_file) as file:
         actual = json.load(file)
@@ -402,7 +407,7 @@ def test_download_playlists_with_youtube_url_with_youtube_url_cache(monkeypatch,
 
     monkeypatch.setattr(ytmusic.YTM, "search_youtube_music", mocked_func)
 
-    playlist.download_playlists(args_list, sptfy_mock)
+    commands.download(args_list, sptfy_mock)
 
     with open(temp_file) as file:
         actual = json.load(file)
@@ -423,7 +428,7 @@ def test_download_playlists_with_youtube_url_with_preloaded_cache(monkeypatch, s
 
     monkeypatch.setattr(ytmusic.YTM, "get_youtube_url", lambda self, song: "https://music.youtube.com/watch?v=wxyz")
 
-    playlist.download_playlists(args_list, sptfy_mock)
+    commands.download(args_list, sptfy_mock)
 
     with open(temp_file) as file:
         actual = json.load(file)
@@ -439,7 +444,7 @@ def test_preload_youtube_url_cache_file_not_found(ytm_mock):
     expected_exit_code = 1
 
     with pytest.raises(SystemExit) as exit_wrapper:
-        playlist.preload_youtube_url_cache(ytm_mock, test_case, False)
+        preload_youtube_url_cache(ytm_mock, test_case, False)
     assert exit_wrapper.type == SystemExit
     assert exit_wrapper.value.code == expected_exit_code
 
@@ -450,7 +455,7 @@ def test_preload_youtube_url_cache(ytm_mock):
         "https://example.invalid": "https://music.youtube.com/watch?v=wxyz",
     }
 
-    playlist.preload_youtube_url_cache(ytm_mock, test_case, False)
+    preload_youtube_url_cache(ytm_mock, test_case, False)
     assert ytm_mock.cache == expected
 
 
@@ -461,5 +466,5 @@ def test_preload_youtube_url_cache_with_unvalidated_urls(ytm_mock):
         "https://example2.invalid": "https://music.youtube.com/watch?v=1234",
     }
 
-    playlist.preload_youtube_url_cache(ytm_mock, test_case, True)
+    preload_youtube_url_cache(ytm_mock, test_case, True)
     assert ytm_mock.cache == expected
